Update increment rounding functionality (#53)
`NormalizedTimeDuration` needs to be a `i128` and not a `f64` for safety
as well as the new duration rounding algoritms. This means we need to be
able to round both `i128` and `f64` for the library. So this commit
currently updates the approach to rounding to allow handling of `i128`
and `f64` at the same time, which allows `NormalizedTimeDuration` to be
an `i128`.