Refactor `TemporalFields` into `CalendarFields` (#95)
This PR refactors `TemporalFields` into `CalendarFields`. I'm leaving it
as a draft for now as I want to run plug this into Boa and run some
tests beforehand.

General Context:

This builds on the work from #89 and #92.

With the partials implemented for the `with` methods. The values that
would be provided can now be accounted for. Primarily, the time fields
really didn't need to exist in the `TemporalFields` representation
anymore (along with the TimeZone and offset which will probably be
living on the `PartialZonedDateTime`). Furthermore, the partials were
now handling the `undefined` case of the fields, so the calendar
specific fields no longer needed to be an `Option`. In general, I also
just think that this is a better native representation of the fields
values in comparison to the previous version.

Furthermore, this PR also implements better handling for Era and Month
Codes according to the [Intl era and monthCode
proposal](https://tc39.es/proposal-intl-era-monthcode/#sec-temporal-isvalidmonthecodeforcalendar).

There are a couple points where feedback/bike-shedding would definitely
be appreciated. I think the method names in `calendar_types.rs` are
getting a bit unyielding. I am also a bit worried about introducing the
`CalendarMethods` trait, but I do think it makes the ergonomics around
passing `Date`, `DateTime`, and, eventually, `ZonedDateTime` as a
fallback convenient.

cc: @sffc