Move to no_std (#101)
This moves the crate over to `no_std` so it can be used on more
platforms and doesn't require linking in the rust standard library.

I wasn't sure what to do with `temporal_assert`. I think in general the
printing functionality of `temporal_assert` should be behind some
feature anyway? I'm not a huge fan of libraries choosing to output stuff
to stdout. Perhaps the log crate can be used as an optional dep instead.