Adjust `RelativeTo` according to specification and implementation (#140)
These are adjustments to the `RelativeTo` option based off integrating
the most recent commits into Boa.

The primary issue is that `RelativeTo` is a specific `to_relative_to`
abstract operation (specifically,
[GetTemporalRelativeToOption](https://tc39.es/proposal-temporal/#sec-temporal-gettemporalrelativetooption))
that parses a string or property bag obj as a `ZonedDateTime` but has
special cases to fall back to a `PlainDate` where that is possible.

This also removes keeping `RelativeTo` as a reference, because it's not
really possible to have a reference where the `PlainDateTime`.