Extend implementation of `to_ixdtf_string` to more types (#155)
This PR continues the work on implementing `toString` functionality in
`temporal_rs`.

It completes the following:

- Adds `to_ixdtf_string` functions on `DateTime`, `Time`,
`ZonedDateTime`, and `Instant`.
- Adds `ToStringRoundingOptions` options type.
- Adds an `IxdtfStringBuilder` struct to assist with building `ixdtf`
strings.

So far I have verfied that the PR has a 100% conformance rate for
PlainDateTime. I plan to test the others, but any poentially issues that
may arise with those could always be completed in subsequent PRs.