Implement Neri-Schneider calculations (#147)
This PR changes the current date calculations to one's layed out by
[Neri-Schneider's Euclidean affine functions
paper](https://onlinelibrary.wiley.com/doi/full/10.1002/spe.3172).

There is still some documentation that needs to be completed before this
is ready to merge. I was also hoping to double check the assembly but
I'm sort of viewing that as part of the documentation. In the meantime,
I'm posting this as a draft for feedback.

---------

Co-authored-by: Cassio Neri <cassio.neri@gmail.com>