More FFI APIs (#178)
Unfortunately the `temporal_core` stuff breaks all the duration APIs,
and I don't understand what it's doing. I wrote this PR before that
change, help updating it to work with Duration appreciated.

In general I don't think it's a good idea to have two separate Duration
types with differing APIs. Opting in to features might enable new APIs,
but it's not kosher to disable APIs that way, which seems to be the case
here.