Build out stubs for remaining unimplemented methods (#202)
Unimplemented methods not being stubbed came up today in the ICU4X call.
This implements stubs for the remaining unimplemented methods that I'm
currently aware of nobody working on (in order to minimize any conflicts
with ongoing work).

This PR also takes a stab at adding the `temporal_capi` methods.

CC: @Manishearth