Reject datetime when fraction digits are too large (#229)
We were unwrapping rather than rejecting on the new `to_nanosecond`
method for `Fraction`.

Context:
`to_nanosecond` returns an `Option<u32>`, which represents the
nanoseconds value when the digits are < 9, but returns `None` if the
digits exceed 9.

I think this mostly points towards a doc improvement needed in `ixdtf`
that also links to `to_truncated_nanosecond` to show the difference
between the two cases.