Clean up API ergonomics for calendar methods (#235)
This PR removes a lot of the `TemporalResult` returns from most of the
calendar methods.

This is made possible on most methods because we no longer need to
support Not Yet Implemented errors.

There is also an expect added in the conversion from `IsoDate` ->
`icu_calendar::Date<Iso>`. This is tested to be valid in a test added to
`iso.rs` for the max bounds of `IsoDate`.

The core changes in this PR are made to `iso.rs` and `calendar.rs` by
extension. The rest is mostly updating all the calendar methods
accordingly.

Note: calendar methods that are still not fully implemented across
various calendars were left as a `TemporalResult`