Rename methods on `Now` and add a test (#243)
This PR adjusts the `Now` core API and adds a test using mock system
data.

The two goals of this PR are the API and implementing at least one more
`Now` test. As far as the API goal, this PR attempts to make the API a
bit more explicit and clear. If there's any suggestion for a better
method name for the core methods, please feel free to provide feedback.

In the future, we should could add more now unit tests using mock data
to ensure the behavior is stable and as expected.