Implement IANA normalizer baked data provider (#251)
This PR begins the initial work for implementing baked data providers.

There is probably going to be a lot more work to be done on this topic
overall (especially for tzif support). This is meant as a minimum proof
of concept that adds support for IANA identifier normalization, and adds
that support to `FsTzdbProvider`.

List of changes:

- Adds bakeddata internal tool (this will need to be iterated on and
further built out probably).
- Adds a `temporal_provider` crate for sourcing zoneinfo data and
bakeddata struct definitions.
- Adds the newly baked `SINGLETON_IANA_NORMALIZER` to `FsTzdbProvider`

General points of consideration:

- I'm not entirely sure how much of an infrastructure we want to build
out to begin with, so the goal here was to be fairly minimal in the
approach.
- My assumption is that the large diff here is from adding in the
zoneinfo files into the repository. We can long term request the data
from IANA as they are publicly hosted, but I do think that the locality
of having the files in the project for reference is nice. There is a
good argument for removing the makefile, newsfile, and a couple things
like that.
- For bakeddata's output, I've added both a debug and file path. I think
this is probably going to be a good approach for maintenance long term.
The thought approach here is that we should ideally not be committing
generated binary without having a debug view of it.