Remove the Temporal prefix from Unit, RoundingMode, and UnsignedRoundingMode (#254)
This PR closes #222.

The background was mentioned in the initial issue. But to restate for
this PR, `TemporalUnit`, `TemporalRoundingMode`, and
`TemporalUnsignedRoundingMode` come from the very early days of
temporal_rs, dating alway back to the initial merge into Boa. The
explicit name was to prevent any clashing with `Intl`. With temporal_rs
pulled out of Boa and standing more formally as it's own crate, it's
long been overdue to remove the prefix as it is redundant.

The general proposal is to change the below names:

  - `temporal_rs::options::TemporalUnit` -> `temporal_rs::options::Unit`
- `temporal_rs::options::TemporalRoundingMode` ->
`temporal_rs::options::RoundingMode`
- `temporal_rs::options::TemporalUnsignedRoundingMode` ->
`temporal_rs::options::UnsignedRoundingMode`
 
CC: @Manishearth 

The change in this PR also currently extends to `temporal_capi`. Would
there be a preference around keeping the more explicit temporal prefix,
or would there not be much of an effect either way in your opinion?