Rename the provider crate (#289)
Related to the ongoing discussion in #277.

This PR updates the provider crate to `timezone_provider`, which seemed
to be a general agreed upon name.

So far it does not make any major decision on what is currently in the
crate. The primary goal is to have the crate name in aligned with a go
forward value so as to not block a new release.