Implement a builder API for Now (#296)
This PR implements a builder API for now and closes #249.

It has some notable changes to the approach to `Now`.

Mainly `Now` is essentially cannot be built without using the new
`NowBuilder` and any method called on `Now` consumes it.

For general purpose, a `Temporal` namespace struct was created with
`now` and `try_now` methods, which uses the `NowBuilder` to construct a
`Now` with values from the system.

This does differ a bit from the specification in that we are eagerly
calling the host system time zone, rather than lazily when the user time
zone is `None`. But I think this can still be worked around on the
engine side if lazily calling is preferred and is more a place where the
native Rust API will differ from the engine implementation. The work
around to this is a trait based approach that is available to see on the
`now-builder` branch, but I think this a simpler and cleaner approach
than leaning on traits.

List of changes:

- `Now` updated and `NowBuilder` added.
- Feature flagged `Temporal` struct
- `time` module renamed to `unix_time`