Add ZonedDateTime FFI (#329)
<s>Based on https://github.com/boa-dev/temporal/pull/326</s>
<s>Based on https://github.com/boa-dev/temporal/pull/327</s>

Fills in FFI for ZDT. Where there are missing provider methods I added
some.

I was hoping to tackle ZDT after the provider stuff is ready, but I
decided instead to have the full API surface so I can fill everything in
on V8, and we can revisit things later. I have deliverately attempted to
design the FFI API such that it is overall immune to provider changes.

One major difference from the temporal_rs ZDT API is that most of the
getters are infallible and have GIGO behavior on errors. See
https://github.com/boa-dev/temporal/issues/328