Add ErrorMessage enum, start using it (#355)
This is a step towards #113. I didn't finish converting all of the
enums, just some of them to give an idea of what this might look like.

This error enum will be private for now: It's primarily encompassing
messages, and will wrap all *static* string messages. The non-static
ones I think we should convert to something else at the boundary,
potentially logging them with a feature.


Once this is fully filled in, I plan to:

 - Replace the .msg field with the enum
- Add to/from integer methods for the enum, exposed as helpers on
TemporalError
 - Use the integer value in FFI TemporalError.


This is not primarily designed with user matching in mind: it can be
matched by the user, but I don't think this is the right level of
granularity. Instead, we can use this error message ontology to map to a
more structured public error ontology if we want. Basically, in the
medium term we can have a RustErrorKind that can be internally obtained
from ErrorMessage, which we expose to the user.


Thoughts?