Implement zoneinfo parsing/compilation and add TZif structs (#257)
It's time to finally tackle the zoneinfo compilation and tzif sourcing
issue. The work here lays the baseline for to handle sourcing. There's
still a lot more to add, and maybe some bugs to uncover. But this is
also major progress to supporting the ability to source tzifs.

Opening as a draft first for early feedback due to just how dense this
actually is. So far, the work here allows us to parse zoneinfo files and
compile sets of transition data from the parsed zoneinfo files. There is
also related work in provider to setup the tzif structures.

The steps missing from this PR is:
- [x] All that's left is to bridge the gap from transition sets ->
TZIFs,
- [x] Plug everything into the bakeddata tool or a new connected tool.

---------

Co-authored-by: Sean Burke <leftmostcat@gmail.com>