Fix float precision check (#457)
This is a workaround for #189.

Most of the time, storing the duration as floats or integers doesn't
matter: all valid durations are in the range of "safe" floats where they
can be represented as distinct integers.

However, the check for "safe" floats itself needs to work with float
values, or at least numerical values that have been normalized to their
floating point equivalent.

This casts the days-and-lower values to/from f64 to round them as
necessary.