Fix behavior of ns-to-seconds casting for negative values (#475)
In a couple cases, casting inexact ns to seconds for positive values
doesn't change much, since the transition starts on a whole second
offset anyway.

But when doing so for negative values, you risk working with the wrong
seconds value. This leads to buggy behavior for times like
`1968-02-18T01:59:59.999999999+00:00[Europe/London]` (one ns before a
transition).

I think the best way to clean this type of thing up is to have a
`NanosecondsAsSeconds` value that is just an i128 wrapper but has
Ord/etc implementations against Seconds. But it's tricky, and I want to
have the carefully written correct behavior first.