Reuse ZDT parsing logic from PartialZDT (#483)
Progress on #408

I don't like having two very similar pieces of code; especially since
all of our tests use `ZDT::from_utf8` but implementors will have to use
`PartialZDT::try_from_utf8` to deal with #408.

....and I switched things over and immediately discovered a problem in
that `START-OF-DAY` wasn't representable in `PartialZDT`. Good thing I
did; this may have implications down the line for the rest of #408
(though by and large `START-OF-DAY` is only important for ZDT, it seems)


As I work on #408, I plan to do the same for all of the other callers.
This also helps ensure that I won't break any tests in implementors
using Partial types for parsing.