Fix broken tests + update various abstract operations
Many broken tests were related to DateDuration needing
to be an intermediate computational version. For instance,
days can exceed the normal threshold in some calculations.

There was also a number of operations that needed to be
updated the the current specification in order to fix the
bugs that came from this current update. Amongst them were
nudge_to_calendar_unit, nudge_to_day_or_time, various
InternalDuration record operations.

It's worth noting that this current update preserves
the old "normalized" specification language. Those should
probably be updated in follow ups to the newer language
for specification consistency, but that seemed to be out
of scope for the current PR.