Add separate parsed variants of all types (#486)
This adds ParsedFoo types to be used by JS engines 

I was initially going to clean up PartialZDT as well, and return it to
its original purpose, but I'll let @nekevss mess with that code.

Note that we do not need a ParsedTime: parsing and validation is
performed concurrently for time parsing.

Fixes #408

---------

Co-authored-by: Kevin Ness <46825870+nekevss@users.noreply.github.com>