Add `CalendarFields` concept and integate it with API and Partial structs (#487)
Addresses issue #349

This PR introduces a `CalendarFields` concept to deal with the
difference between the `from_partial` and `with` API.

It adds a number of "fields" types: `CalendarFields`,
`YearMonthCalendarFields`, `DateTimeFields`, `ZonedDateTimeFields`, and
updates: `PartialDate`, `PartialYearMonth`, `PartialDateTime`, and
`PartialZonedDateTime`.

Most of the general changes are to `from_partial` and `with` APIs with
some secondary changes occurring wherever `PartialDate` was being used
before hand and could now use `CalendarFields` instead.

There's probably a secondary discussion to be had about what the general
API for the Partial objects and the Fields object. I tried to avoid that
in this PR as much as possible (outside of adding a little bit of
convenience for now) to defer it until there was some feedback.