Further normalize MWDs, add Vancouver and Santiago tests (#504)
We were not handling the fact that the "first Friday of the month" is
not necessarily after the "first sunday of the month".

There's a case to be made here that this logic isn't worth it and we
should just normalize the Mwds to `Seconds` instead of trying to turn
them into something comparable. I don't have a strong opinion there,
someone who wants to look into that can benchmark it.

This also adds some tests for America/Santiago, which has some fun
quirks I discovered.