Make errors Copy (#528)
This implements the strategy somewhat discussed in
https://github.com/unicode-org/icu4x/issues/6904.

Previously, all allocated error messages from Temporal would not carry
over FFI since we were reluctant to make the error type need a
destructor.

TemporalError is now fully Copy, and can contain ICU4X/ixdtf errors. It
is still converted to a string over FFI, which does not allocate. This
leads to a minimal loss of information.