Update zoneinfo compilation in zoneinfo crate (#532)
This PR updates the internal approach to compiling zoneinfo files into
transition data.

The main goal of this PR is to support slim format compiled data in
`timezone_provider`, which this PR should unblock as a
`CompiledTransition` is now compiled in slim by default. This new
approach also has some added benefits: one, it appears to be more
correct as it fixes previously known bugs that needed to be addressed,
and two, this approach is simpler to understand and doesn't do some of
the backtracking that was occurring in the old approach.