Add provider APIs to capi (#544)
The general design: all compiled_data methods get with_provider
versions; the compiled_data call is tweaked to use a compiled provider.
In theory I could choose to not proxy compiled_data methods to
with_provider ones but it's cleaner this way, and as long as the build
configuration doesn't enable two backends it should be zero-cost.

Haven't done everything yet, but this is the approach.