Refactor Now to be trait based for lazy host getters (#560)
Closes #547

This basically removes the NowBuilder approach for a trait based
approach.

In order to do this, it adds the HostClock, HostTimeZone, and HostHooks
traits. I could see adding the HostHooks trait as potentially
superfluous / controversial with the other two traits already present,
but I thought it may provide a bit cleaner interface. Let me know what
you think. Definitely willing to work shop this one a bit more if need
be.

I did test this in Boa locally and there shouldn't be any regressions,
which is nice.