Bump v0.1.0 release for temporal_rs and temporal_capi (#589)
Drafting a pull request for the v0.1 release.

Merging is going to be dependent on two questions / tasks:

- [x] What should the version for `timezone_provider` and `zoneinfo_rs`
be?
- [ ] Is the [release
post](https://github.com/boa-dev/boa-dev.github.io/pull/210) reviewed
and ready?

Regarding the question around versioning, I've left them still in the
0.0.x path until the APIs can be a bit more stabilized. But I'm open to
adding them to a 0.1.0 version. My main concern is if we do make a
breaking change to `timezone_provider` in the future (like stablizing
the ZeroCompiledTzdbProvider) , then we may have to adjust the versions
then for `timezone_provider` to be a 0.2.0 and `temporal_rs` to be a
0.1.1. I'd prefer to do a delayed 0.1 for those crates, but that is not
an incredibly strong position.