Remove HijriSimulatedMecca / islamic-rgsa (#597)
The spec doesn't have it anymore.

We use AnyCalendarKind parsing code so we still parse it, we just turn
it into Iso.

I don't quite know the best way to handle this for temporal_rs:
Calendar::new is infallible, but it's mostly internally used in parsing,
and we accept AnyCalendarKinds in our public API.

I think in the long run we should switch to `try_new` and in the IXDTF
parsing case have a fallback to Iso.