remove unneeded clone in TzdbResolver::get (#628)
I noticed that every time when a timezone is accessed via `ResolvedId`
in the cache, the full `Tzif` struct is cloned, which contains multiple
allocations.

Just from poking around in the code, this seems to happen quite
frequently, for example when a new `ZonedDateTime` is created.


In the future the dx can be improved once `RwLockReadGuard::map` is
stabilized, see rust-lang/rust#117108, though I think it will take some
time.