Correct sign in nudge_to_zoned_time
According to the spec, the sign argument to NudgeToZonedTime should
be either 1 or -1, but in temporal_rs, the Sign type can also represent
the 0 sign. Add code to explicitly handle the zero sign case when rounding
in nudge_to_zoned_time. This was causing a wrong result when rounding
the zero duration (see https://github.com/boa-dev/temporal/issues/646 ).