Update to new ComputeNudgeWindow spec text (#636)
Fixes #635 


Updates to https://github.com/tc39/proposal-temporal/pull/3172


<s>This passes the test, but it fails other tests. The problem is
probably that IncrementRounder isn't written in terms of r1 and r2.
@nekevss, I don't understand the full purpose of IncrementRounder being
written the way it is: mind having a look?

The new spec text changes how r1 and r2 are computed, and they're passed
down to IncrementRounder in the spec but not the code, so I assume that
IncrementRounder is recomputing them somehow from the inputs.</s>


I wrote this to not use IncrementRounder. I'm not 100% sure if the
implementation is correct when it comes to how it handles mathematical
values.

---------

Co-authored-by: Kevin Ness <nekevss@gmail.com>
Co-authored-by: Kevin Ness <46825870+nekevss@users.noreply.github.com>